feat: 관리자 페이지 디렉토리 구조 설정(레이아웃)
