+++ src/app/(admin)/admin/certificates/page.tsx
@@ -1,0 +1,15 @@
+import { AdminSidebar } from '@/features/admin'
+import { AdminDashboard } from '@/features/admin/components/AdminDashboard'
+
+export default function AdminCertificatePage() {
+  return (
+    <>
+      <div className="custom-container">
+        <div className="custom-card flex gap-6">
+          <AdminSidebar />
+          <AdminDashboard />
+        </div>
+      </div>
+    </>
+  )
+}
