+++ src/app/(admin)/admin/layout.tsx
@@ -1,0 +1,10 @@
+import HeaderSection from '@/components/layout/header/HeaderSection'
+
+export default function AdminLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <>
+      <HeaderSection />
+      {children}
+    </>
+  )
+}
