--- src/app/admin/page.tsx
@@ -1,5 +1,0 @@
-import { AdminDashboard } from '@/features/admin'
-
-export default function AdminPage() {
-  return <AdminDashboard />
-}
